+++ ec/__main__.py
@@ -29,23 +29,41 @@ def main():
       core.start(load_module(target_path), argv)
       
     elif path.isdir(target_path): # launch the dir with its children as groups
+      import imp
       from glob import glob
+      from collections import OrderedDict
+      
       from modules.classes import Group
       
+      Container = imp.new_module(target_path)
+      Members = OrderedDict()
+      
+      for Module in [load_module(module_path) for module_path in glob('%s/*.py' % path.abspath(target_path)) if path.isfile(module_path)]:
+        __ec_member__ = getattr(Module, '__ec_member__', None)
+        if __ec_member__: # register only modules that are designed for ec
+          __ec_member__.__load_members__()
+          name = Module.__name__
+          print Module, name, '--------'
+          setattr(Container, name, Module)
+          Members[name] = __ec_member__
+      
+      BaseGroup = Group(Container, {'name': target_path}) # brand the Container
+      BaseGroup.Config['Members'] = Members
+      
+      '''
       class Container: # a class to emulate a module
         def __init__(self, Modules):
           for Module in Modules:
-            setattr(self, Module.__name__, Module)
-        
-      Container = Container([load_module(name) for name in glob('%s/*.py' % path.abspath(target_path)) if path.isfile(name)])
+            if hasattr(Module, '__ec_member__'): # register only modules that are designed for ec
+              setattr(self, Module.__name__, Module)
+      Container = Container()
+      '''
       
-      Group(Container, {'name': target_path}) # brand the Container
-      
+      BaseGroup.__load_members__()
       core.start(Container, argv)
       
     else:
-      show_usage()
-  
+      show_usage()  
 
 def show_usage():
   print 'Usage:\n\t$ ec script/dir [flag] [command] [args]'