+++ tests/support/nester.py
@@ -1,0 +1,12 @@
+"""
+nester
+======
+
+Used to test nested modules.
+"""
+from ec.ec import group
+
+import target_script
+
+group(target_script)
+